try catch added to DocumentsController

+++ src/documents/documentsController.ts
@@ -1,4 +1,3 @@
-import { strict } from 'assert';
 import { NextFunction, Request, Response } from 'express'
 import multer from "multer";
 import path from 'path';
@@ -20,16 +19,19 @@ export const upload = multer({ storage: storage })
 export const testPostDocument = (req: Request, res: Response, next: NextFunction) => {
     try {
         const docName = req.file?.filename;
-        const db = new sqlite3.Database('./uploadedFiles.db');
-        const query = `INSERT INTO documents(name) VALUES ('${docName}')`;
-        db.run(query);
-        db.close();
-        res.status(200).send('Your file is successfully uploaded!');
+        if (typeof docName !== 'string') {
+            throw new Error("Incorrect document name")
+        } else {
+            const db = new sqlite3.Database('./uploadedFiles.db');
+            const query = `INSERT INTO documents(name) VALUES ('${docName}')`;
+            db.run(query);
+            db.close();
+            res.status(200).send('The file is successfully uploaded!');
+        }
     } catch (error) {
-        next(alert("Sorry, couldn't upload the document :("))
-        // res.status(404).send(`Oops! There is something wrong: ${error}`);
+        res.status(400).send(`Oops! ${error}`);
     }
-    
+
 };
 
 export const getAllFiles = (req: Request, res: Response) => {
@@ -50,8 +52,7 @@ export const getAllFiles = (req: Request, res: Response) => {
         });
         db.close();
     } catch (error) {
-        alert("Sorry, couldn't download your documents :(")
-        res.status(404).send(`Oops! There is something wrong: ${error}`);
+        res.status(400).send(`Oops! ${error}`);
     }
 };
 
@@ -59,31 +60,32 @@ export const deleteDocument = async (req: Request, res: Response, next: NextFunc
     try {
         const docId = req.params.id;
         const docName = req.params.name;
+        if (docId === undefined || docName === undefined) {
+            throw new Error('docId or docName is undefined')
+        }
+
         const db = new sqlite3.Database('./uploadedFiles.db');
         const sql = `DELETE FROM documents WHERE id = ${docId}`;
         db.run(sql);
         db.close();
 
-        if (typeof docName !== 'string') {
-            alert("Sorry, couldn't delete the document :(");
-        } else {
-            const pathToDoc = path.join("public", "files", docName)
-            await fs.promises.unlink(pathToDoc);
-        };
-         res.status(200).send('Document deleted successfully');
+        const pathToDoc = path.join("public", "files", docName)
+        await fs.promises.unlink(pathToDoc);
+        res.status(200).send('Document deleted successfully');
     } catch (error) {
-        next(alert("Sorry, couldn't delete the document :("));
-        // res.status(404).send(`Oops! There is something wrong: ${error}`);
+        res.status(400).send(`Oops! ${error}`);
     }
 }
 
 export const downloadDocument = async (req: Request, res: Response) => {
     try {
         const docName = req.params.name;
+        if (typeof docName !== 'string') {
+            throw new Error("Document name is not astring")
+        };
         const pathToDoc = path.join("public", "files", docName)
         res.download(pathToDoc);
     } catch (error) {
-        alert("Sorry, can't download the document :(")
-        res.status(404).send(`Oops! There is something wrong: ${error}`);
+        res.status(400).send(`Oops! ${error}`);
     }
 }
